Bugfix/remove deb uris

+++ ubuntu-starter.sh
@@ -8,61 +8,61 @@ update_and_upgrade() {
 }
 
 create_default_aliases() {
-    # Delete all local branches except master
-    echo "alias dlb=\"git branch | grep -v 'master' | xargs git branch -D\"" >> ~/.bashrc
-
-    # Does the same if zsh is installed
-    [ -f ~/.zshrc ] && echo "alias dlb=\"git branch | grep -v 'master' | xargs git branch -D\"" >> ~/.zshrc
+    if [ `alias | grep dlb | wc -l` > 0 ]; then
+        # Delete all local branches except master
+        [ -f ~/.zshrc ] && echo "alias dlb=\"git branch | grep -v 'master' | xargs git branch -D\"" >> ~/.zshrc
+    fi
 }
 
 package_exists() {
-	return `dpkg-query -l | grep $1 | wc -l`
+    dpkg -s "$1" >/dev/null 2>&1
 }
 
 
 # Upgrading system before anything
 echo ">>> Upating system..."
+
 update_and_upgrade
 
 
 # apt packages
-echo ">>> Installing new packages..."
+echo ">>> Looking for new packages..."
 
 for pkg in `cat packages.apt`; do
-	if [ ! "package_exists $pkg" ]; then
-		sudo apt install $pkg -y
-	fi
+    if package_exists $pkg; then
+        echo "$pkg is already installed."
+    else
+        sudo apt-get install $pkg
+    fi
 done
 
 
 # Snap Packages
-echo ">>> Installing snap packages..."
+echo ">>> Looking for new snap packages..."
 
 for snap in `cat packages.snap`; do
-    if [ ! "package_exists $snap" ]; then
-		sudo snap install $snap --classic
+    if package_exists $snap; then
+        echo "$snap is already installed."
+    else
+		sudo snap install $snap
 	fi
 done
 
 
-# deb packages
-echo ">>> Installing deb packages..."
-
-for uri in `cat deb_uris`; do
-    curl -sL -o/var/cache/apt/archives/package.deb $uri && sudo dpkg -i /var/cache/apt/archives/package.deb
-done
-
-
 # SSH Setup
-ssh-keygen
-eval `ssh-agent`
-ssh-add ~/.ssh/id_rsa
+if [ ! -f ~/.ssh/id_rsa.pub ]; then
+    echo ">>> Creating SSH Keys"
+    ssh-keygen
+    eval `ssh-agent`
+    ssh-add ~/.ssh/id_rsa
+fi
 
 
 # OhMyZsh
-echo ">>> Installing OhMyZsh"
-
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+if [ ! -f ~/.zshrc ]; then
+    echo ">>> Installing OhMyZsh"
+    sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+fi
 
 
 # Aliases
@@ -70,9 +70,11 @@ create_default_aliases
 
 
 # Google Chrome
-echo ">>> Installing Google Chrome"
-
-if [ ! "package_exists google-chrome-stable" ]; then
+# Only base64
+if package_exists google-chrome-stable; then
+    echo "Google Chrome is already installed."
+else
+    echo ">>> Installing Google Chrome"
 	wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
 	sudo sh -c 'echo "deb [arch=amd64] https://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google.list'
 	sudo apt-get update
